Fix with query return value (#8)
A `WithQuery should return the same `Returning` type as the query created by the callback. Instead, it was always creating a `SelectQuery` regardless of the type of query that the callback creates. This would end up returning `[]` when with is used on an `UpdateQuery` or `DeleteQuery`, rather than the number of rows affected. This fixes it to return a copy of the same query that is passed in.