Values list: Add test case for extra parameters (#14)
The `values` function can receive values that have extra properties that the definition does not specify.  It is convenient to support this, because a developer may be working with the results from a query that selects additional columns that are not needed for the query that uses the values list.

This adds a test to ensure that only properties from the definition are tokenized and end up in the resulting SQL.  The test passed without any changes needed, just want to make sure that we continue to support this.