Add lateral joins (#17)
This adds lateral join support.

Typically, a lateral join will nest the join conditions inside of the subquery, with an `ON true` clause for the join.  This also adds support for using a boolean with an `ON` clause to handle this use case.

This also adds the ability to give a select query an alias, which converts it into a from item that can be used in other parts of the query with a new name.

`LATERAL` can also be used in a from item, but mammoth currently only supports a single from item, so it wouldn't be useful to support there.  We can re-asses adding `LATERAL` support in from items if we also add multiple from item support.