CI: Ensure Mammoth version in "package.json" matches "package-lock.json" (#43)
When we bump the version number in "package.json", it's common to forget
to run `npm install`, which means "package-lock.json" will not have the
updated version number. For some reason, `npm ci` doesn't check that.

I added a `jq`+`bash` snippet to the GitHub Actions workflow to check
the version.